+++ src/internal/interpreter/map-interpreter.test.ts
@@ -1,4 +1,5 @@
 import { MapDocumentNode, MapHeaderNode } from '@superfaceai/ast';
+import { parseMap, Source } from '@superfaceai/parser';
 import { getLocal } from 'mockttp';
 
 import { CrossFetch } from '../../lib/fetch';
@@ -21,8 +22,11 @@ const header: MapHeaderNode = {
 };
 
 describe('MapInterpreter', () => {
+  let serviceBaseUrl: string;
+
   beforeEach(async () => {
     await mockServer.start();
+    serviceBaseUrl = mockServer.url;
   });
 
   afterEach(async () => {
@@ -366,7 +370,8 @@ describe('MapInterpreter', () => {
   });
 
   it('should call an API', async () => {
-    await mockServer.get('/twelve').thenJson(
+    const url = '/twelve';
+    await mockServer.get(url).thenJson(
       200,
       { data: 12 },
       {
@@ -374,11 +379,11 @@ describe('MapInterpreter', () => {
         'Content-Language': 'en-US, en-CA',
       }
     );
-    const url = mockServer.urlFor('/twelve');
     const interpreter = new MapInterpreter(
       {
         usecase: 'Test',
         security: [],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -518,13 +523,14 @@ describe('MapInterpreter', () => {
   });
 
   it('should call an API with path parameters', async () => {
-    await mockServer.get('/twelve/2').thenJson(200, { data: 144 });
-    const url = mockServer.urlFor('/twelve');
+    const url = '/twelve';
+    await mockServer.get(url + '/2').thenJson(200, { data: 144 });
     const interpreter = new MapInterpreter(
       {
         usecase: 'Test',
         input: { page: '2' },
         security: [],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -604,16 +610,17 @@ describe('MapInterpreter', () => {
   });
 
   it('should call an API with parameters', async () => {
+    const url = '/twelve';
     await mockServer
-      .get('/twelve')
+      .get(url)
       .withQuery({ page: 2 })
       .thenJson(200, { data: 144 });
-    const url = mockServer.urlFor('/twelve');
     const interpreter = new MapInterpreter(
       {
         usecase: 'Test',
         input: { page: 2 },
         security: [],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -693,16 +700,17 @@ describe('MapInterpreter', () => {
   });
 
   it('should call an API with parameters and POST request', async () => {
+    const url = '/checkBody';
     await mockServer
-      .post('/checkBody')
+      .post(url)
       .withJsonBody({ anArray: [1, 2, 3] })
       .withHeaders({ someheader: 'hello' })
       .thenJson(201, { bodyOk: true, headerOk: true });
-    const url = mockServer.urlFor('/checkBody');
     const interpreter = new MapInterpreter(
       {
         usecase: 'Test',
         security: [],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -783,16 +791,15 @@ describe('MapInterpreter', () => {
   });
 
   it('should run multi step operation', async () => {
-    await mockServer
-      .get('/first')
-      .thenJson(200, { firstStep: { someVar: 12 } });
-    await mockServer.get('/second').thenJson(200, { secondStep: 5 });
-    const url1 = mockServer.urlFor('/first');
-    const url2 = mockServer.urlFor('/second');
+    const url1 = '/first';
+    const url2 = '/second';
+    await mockServer.get(url1).thenJson(200, { firstStep: { someVar: 12 } });
+    await mockServer.get(url2).thenJson(200, { secondStep: 5 });
     const interpreter = new MapInterpreter(
       {
         usecase: 'Test',
         security: [],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -882,11 +889,11 @@ describe('MapInterpreter', () => {
   });
 
   it('should call an API with Basic auth', async () => {
+    const url = '/basic';
     await mockServer
-      .get('/basic')
+      .get(url)
       .withHeaders({ Authorization: 'Basic bmFtZTpwYXNzd29yZA==' })
       .thenJson(200, { data: 12 });
-    const url = mockServer.urlFor('/basic');
     const interpreter = new MapInterpreter(
       {
         usecase: 'testCase',
@@ -899,6 +906,7 @@ describe('MapInterpreter', () => {
             password: 'password',
           },
         ],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -946,11 +954,11 @@ describe('MapInterpreter', () => {
   });
 
   it('should call an API with Bearer auth', async () => {
+    const url = '/bearer';
     await mockServer
-      .get('/bearer')
+      .get(url)
       .withHeaders({ Authorization: 'Bearer SuperSecret' })
       .thenJson(200, { data: 12 });
-    const url = mockServer.urlFor('/bearer');
     const interpreter = new MapInterpreter(
       {
         usecase: 'testCase',
@@ -962,6 +970,7 @@ describe('MapInterpreter', () => {
             token: 'SuperSecret',
           },
         ],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -1009,11 +1018,11 @@ describe('MapInterpreter', () => {
   });
 
   it('should call an API with Apikey auth in header', async () => {
+    const url = '/apikey';
     await mockServer
-      .get('/apikey')
+      .get(url)
       .withHeaders({ Key: 'SuperSecret' })
       .thenJson(200, { data: 12 });
-    const url = mockServer.urlFor('/apikey');
     const interpreter = new MapInterpreter(
       {
         usecase: 'testCase',
@@ -1026,6 +1035,7 @@ describe('MapInterpreter', () => {
             apikey: 'SuperSecret',
           },
         ],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -1074,11 +1084,11 @@ describe('MapInterpreter', () => {
   });
 
   it('should call an API with Apikey auth in query', async () => {
+    const url = '/apikey';
     await mockServer
-      .get('/apikey')
+      .get(url)
       .withQuery({ key: 'SuperSecret' })
       .thenJson(200, { data: 12 });
-    const url = mockServer.urlFor('/apikey');
     const interpreter = new MapInterpreter(
       {
         usecase: 'testCase',
@@ -1091,6 +1101,7 @@ describe('MapInterpreter', () => {
             apikey: 'SuperSecret',
           },
         ],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -1138,13 +1149,15 @@ describe('MapInterpreter', () => {
   });
 
   it('should call an API with multipart/form-data body', async () => {
-    await mockServer.post('/formdata').thenCallback(request => {
+    const url = '/formdata';
+    await mockServer.post(url).thenCallback(async request => {
+      const text = await request.body.getText();
       if (
-        request.body.text &&
-        request.body.text.includes('formData') &&
-        request.body.text.includes('myFormData') &&
-        request.body.text.includes('is') &&
-        request.body.text.includes('present')
+        text &&
+        text.includes('formData') &&
+        text.includes('myFormData') &&
+        text.includes('is') &&
+        text.includes('present')
       ) {
         return {
           json: { data: 12 },
@@ -1154,11 +1167,11 @@ describe('MapInterpreter', () => {
 
       return { json: { failed: true }, statusCode: 400 };
     });
-    const url = mockServer.urlFor('/formdata');
     const interpreter = new MapInterpreter(
       {
         usecase: 'testCase',
         security: [],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -1229,15 +1242,16 @@ describe('MapInterpreter', () => {
   });
 
   it('should call an API with application/x-www-form-urlencoded', async () => {
+    const url = '/urlencoded';
     await mockServer
-      .post('/urlencoded')
+      .post(url)
       .withForm({ form: 'is', o: 'k' })
       .thenJson(201, { data: 12 });
-    const url = mockServer.urlFor('/urlencoded');
     const interpreter = new MapInterpreter(
       {
         usecase: 'testCase',
         security: [],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -1475,12 +1489,13 @@ describe('MapInterpreter', () => {
   });
 
   it('should correctly return from operation', async () => {
-    await mockServer.get('/test').thenJson(200, {});
-    const url = mockServer.urlFor('/test');
+    const url = '/test';
+    await mockServer.get(url).thenJson(200, {});
     const interpreter = new MapInterpreter(
       {
         usecase: 'Test',
         security: [],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -1648,94 +1663,9 @@ describe('MapInterpreter', () => {
     expect(result.isOk() && result.value).toEqual({ answer: 42 });
   });
 
-  it('should merge results', async () => {
-    const ast: MapDocumentNode = {
-      kind: 'MapDocument',
-      header,
-      definitions: [
-        {
-          kind: 'MapDefinition',
-          name: 'test',
-          usecaseName: 'test',
-          statements: [
-            {
-              kind: 'CallStatement',
-              operationName: 'foo',
-              arguments: [],
-              statements: [
-                {
-                  kind: 'OutcomeStatement',
-                  isError: false,
-                  terminateFlow: false,
-                  value: {
-                    kind: 'ObjectLiteral',
-                    fields: [
-                      {
-                        kind: 'Assignment',
-                        key: ['a'],
-                        value: {
-                          kind: 'PrimitiveLiteral',
-                          value: 41,
-                        },
-                      },
-                    ],
-                  },
-                },
-              ],
-            },
-            {
-              kind: 'CallStatement',
-              operationName: 'bar',
-              arguments: [],
-              statements: [
-                {
-                  kind: 'OutcomeStatement',
-                  isError: false,
-                  terminateFlow: false,
-                  value: {
-                    kind: 'ObjectLiteral',
-                    fields: [
-                      {
-                        kind: 'Assignment',
-                        key: ['b'],
-                        value: {
-                          kind: 'PrimitiveLiteral',
-                          value: 42,
-                        },
-                      },
-                    ],
-                  },
-                },
-              ],
-            },
-          ],
-        },
-        {
-          kind: 'OperationDefinition',
-          name: 'foo',
-          statements: [],
-        },
-        {
-          kind: 'OperationDefinition',
-          name: 'bar',
-          statements: [],
-        },
-      ],
-    };
-    const interpreter = new MapInterpreter(
-      {
-        usecase: 'test',
-        security: [],
-      },
-      { fetchInstance }
-    );
-    const result = await interpreter.perform(ast);
-    expect(result.isOk() && result.value).toEqual({ a: 41, b: 42 });
-  });
-
   it('should perform operations with correct scoping', async () => {
-    await mockServer.get('/test').thenJson(200, {});
-    const url = mockServer.urlFor('/test');
+    const url = '/test';
+    await mockServer.get(url).thenJson(200, {});
     const ast: MapDocumentNode = {
       kind: 'MapDocument',
       header,
@@ -1883,6 +1813,7 @@ describe('MapInterpreter', () => {
       {
         usecase: 'Test',
         security: [],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -2197,6 +2128,40 @@ describe('MapInterpreter', () => {
     const result = await interpreter.perform(ast);
 
     expect(result.isOk() && result.value).toEqual({ results: ['Z', 'Y', 'X'] });
+  });
+
+  it('should break from iteration', async () => {
+    const ast: MapDocumentNode = parseMap(
+      new Source(
+        `
+        profile = "test@1.0"
+        provider = "test"
+  
+        map Test {
+          letters = ['x', 'y', 'z']
+          results = []
+          call foreach(letter of letters.reverse()) TestOp(letter = letter) {
+            results = results.concat(outcome.data)
+
+            return map result { results = results }
+          }
+        }
+        operation TestOp {
+          return args.letter.toUpperCase()
+        }
+        `
+      )
+    );
+    const interpreter = new MapInterpreter(
+      {
+        usecase: 'Test',
+        security: [],
+      },
+      { fetchInstance }
+    );
+    const result = await interpreter.perform(ast);
+
+    expect(result.isOk() && result.value).toEqual({ results: ['Z'] });
   });
 
   it('should perform an inline iterating call', async () => {
@@ -2544,8 +2509,8 @@ describe('MapInterpreter', () => {
   });
 
   it('should be able to use input in path parameters', async () => {
-    await mockServer.get('/twelve').thenJson(200, { data: 12 });
-    const baseUrl = mockServer.urlFor('/twelve').replace('/twelve', '');
+    const url = '/twelve';
+    await mockServer.get(url).thenJson(200, { data: 12 });
     const ast: MapDocumentNode = {
       kind: 'MapDocument',
       header,
@@ -2558,7 +2523,7 @@ describe('MapInterpreter', () => {
             {
               kind: 'HttpCallStatement',
               method: 'GET',
-              url: `${baseUrl}/{input.test}`,
+              url: `/{input.test}`,
               responseHandlers: [
                 {
                   kind: 'HttpResponseHandler',
@@ -2596,6 +2561,7 @@ describe('MapInterpreter', () => {
         usecase: 'Test',
         input: { test: 'twelve' },
         security: [],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );
@@ -2680,7 +2646,8 @@ describe('MapInterpreter', () => {
   });
 
   it('should make response headers accessible', async () => {
-    await mockServer.get('/twelve').thenJson(
+    const url = '/twelve';
+    await mockServer.get(url).thenJson(
       200,
       {},
       {
@@ -2689,11 +2656,11 @@ describe('MapInterpreter', () => {
         Data: '12',
       }
     );
-    const url = mockServer.urlFor('/twelve');
     const interpreter = new MapInterpreter(
       {
         usecase: 'Test',
         security: [],
+        serviceBaseUrl,
       },
       { fetchInstance }
     );