+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Log warning when user does not handle rejections properly
+- **BREAKING CHANGE**: absolute URLs are no longer allowed in maps
+- **BREAKING CHANGE**: interpolation no longer works in base URLs
+
+### Changed
+- Removed unnecessary result propagation in interpreter - using stackTop.result instead
+- Terminal outcome statements should now correctly exit from iteration
+
+## [0.0.37-beta.0] - 2021-10-05
+### Added
+- Make `configuration` of `BoundProfileProvider` public
 
 ## [0.0.36] - 2021-09-29
 ### Fixed
@@ -243,7 +253,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Profile parameter validator
 - CI/CD workflows
 
-[Unreleased]: https://github.com/superfaceai/one-sdk-js/compare/v0.0.36...HEAD
+[Unreleased]: https://github.com/superfaceai/one-sdk-js/compare/v0.0.37-beta.0...HEAD
+[0.0.37-beta.0]: https://github.com/superfaceai/one-sdk-js/compare/v0.0.36...v0.0.37-beta.0
 [0.0.36]: https://github.com/superfaceai/one-sdk-js/compare/v0.0.36-beta.0...v0.0.36
 [0.0.36-beta.0]: https://github.com/superfaceai/one-sdk-js/compare/v0.0.35-beta.0...v0.0.36-beta.0
 [0.0.35-beta.0]: https://github.com/superfaceai/one-sdk-js/compare/v0.0.34-beta.1...v0.0.35-beta.0